+++ src/test/java/org/ojbc/mondrian/rest/PreCachedMetadataTest.java
@@ -35,5 +35,5 @@ public class PreCachedMetadataTest extends AbstractMondrianRestControllerTest {
 		assertEquals(1, responseHeaders.size());
 		assertEquals("true", responseHeaders.get(0));
 	}
-
+	
 }